+++ tests/testthat/test-other-features.R
@@ -8,9 +8,22 @@ test_with_dir("Can standardize commands from expr or lang", {
   z <- standardize_command(x)
   w <- standardize_command(x[[1]])
   s <- "{\n f(x + 2) + 2 \n}"
+  debug_char0 <-
   expect_equal(y, s)
   expect_equal(z, s)
   expect_equal(w, s)
+})
+
+test_with_dir("debug_command()", {
+  skip_on_cran()
+  txt <- "    f(x + 2) + 2"
+  txt2 <- "drake::debug_and_run(function() {\n    f(x + 2) + 2\n})"
+  x <- parse(text = txt)
+  out1 <- debug_command(x[[1]])
+  out2 <- debug_command(txt)
+  txt3 <- rlang::expr_text(out1)
+  expect_equal(out2, txt2)
+  expect_equal(out2, txt3)
 })
 
 test_with_dir("build_target() does not need to access cache", {